tests: Move integration tests to 'tests'
Move cilium_integration_test to a new 'tests' directory and split it
into smaller parts. This allows faster test iteration as bazel caches
tests that do not need to be re-run.
