+++ BUILD
@@ -3,11 +3,9 @@ licenses(["notice"])  # Apache 2
 load(
     "@envoy//bazel:envoy_build_system.bzl",
     "envoy_cc_binary",
-    "envoy_cc_test",
 )
 
-exports_files(["linux/bpf_common.h", "linux/bpf.h", "linux/type_mapper.h",
-               "proxylib/libcilium.h", "proxylib/types.h"])
+exports_files(["linux/bpf_common.h", "linux/bpf.h", "linux/type_mapper.h"])
 
 envoy_cc_binary(
     name = "cilium-envoy-deps",
@@ -31,25 +29,6 @@ envoy_cc_binary(
         "//cilium:tls_wrapper_lib",
 
         "@envoy//source/exe:envoy_main_entry_lib",
-    ],
-)
-
-envoy_cc_test(
-    name = "cilium_integration_test",
-    srcs = ["cilium_integration_test.cc"],
-    data = [
-        "cilium_proxy_test.json",
-        "proxylib/libcilium.so",
-        "@envoy//test/config/integration/certs",
-    ],
-    repository = "@envoy",
-    deps = [
-        "//cilium:bpf_metadata_lib",
-        "//cilium:network_filter_lib",
-        "//cilium:l7policy_lib",
-        "//cilium:tls_wrapper_lib",
-        "@envoy//test/integration:http_integration_lib",
-        #"@envoy//source/extensions/transport_sockets/tls:context_config_lib",
     ],
 )
