+++ proxylib/BUILD
@@ -1,0 +1,1 @@
+exports_files(["libcilium.h", "types.h", "libcilium.so"])
